app/pyproject.toml
[tool.poetry]
name = "moonraker-logger"
version = "0.1.0"
description = ""
authors = ["trasba <t.rasbach@gmx.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
websockets = "^12.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
